requirements.txt
# Core
httpx>=0.25.0
python-dotenv>=1.0.0
pillow>=10.0.0

# LangChain & Embeddings
langchain>=0.2.0
langchain-community>=0.2.0
langchain-huggingface>=0.0.3
sentence-transformers>=2.2.0

# Vector Store
chromadb>=0.4.0

# Document Loading
pypdf>=3.0.0

# Web Search
duckduckgo-search>=5.0.0

# UI - Streamlit
streamlit>=1.40.0

# APIs
google-generativeai>=0.8.0
mistralai>=1.0.0
